Refactor tzrange and tzwin to inherit from a central tzrangebase object.
