+++ dateutil/tz/_common.py
@@ -61,9 +61,6 @@ class _tzinfo(tzinfo):
     """
     Base class for all ``dateutil`` ``tzinfo`` objects.
     """
-    
-    def __init__(self, *args, **kwargs):
-        super(_tzinfo, self).__init__(*args, **kwargs)
 
     def is_ambiguous(self, dt):
         """
@@ -172,139 +169,46 @@ class _tzinfo(tzinfo):
         return enfold(dt_wall, fold=_fold)
 
 
-class tzrange(_tzinfo):
+class tzrangebase(_tzinfo):
     """
-    The ``tzrange`` object is a time zone specified by a set of offsets and
-    abbreviations, equivalent to the way the ``TZ`` variable can be specified
-    in POSIX-like systems, but using Python delta objects to specify DST
-    start, end and offsets.
+    This is an abstract base class for time zones represented by an annual
+    transition into and out of DST. Child classes should implement the following
+    methods:
 
-    :param stdabbr:
-        The abbreviation for standard time (e.g. ``'EST'``).
+        * ``__init__(self, *args, **kwargs)``
+        * ``transitions(self, year)`` - this is expected to return a tuple of
+          datetimes representing the DST on and off transitions in standard
+          time.
 
-    :param stdoffset:
-        An integer or :class:`datetime.timedelta` object or equivalent
-        specifying the base offset from UTC.
-
-        If unspecified, +00:00 is used.
-
-    :param dstabbr:
-        The abbreviation for DST / "Summer" time (e.g. ``'EDT'``).
-
-        If specified, with no other DST information, DST is assumed to occur
-        and the default behavior or ``dstoffset``, ``start`` and ``end`` is
-        used. If unspecified and no other DST information is specified, it
-        is assumed that this zone has no DST.
-
-        If this is unspecified and other DST information is *is* specified,
-        DST occurs in the zone but the time zone abbreviation is left
-        unchanged.
-
-    :param dstoffset:
-        A an integer or :class:`datetime.timedelta` object or equivalent
-        specifying the UTC offset during DST. If unspecified and any other DST
-        information is specified, it is assumed to be the STD offset +1 hour.
-
-    :param start:
-        A :class:`relativedelta.relativedelta` object or equivalent specifying
-        the time and time of year that daylight savings time starts. To specify,
-        for example, that DST starts at 2AM on the 2nd Sunday in March, pass:
-
-            ``relativedelta(hours=2, month=3, day=1, weekday=SU(+2))``
-
-        If unspecified and any other DST information is specified, the default
-        value is 2 AM on the first Sunday in April.
-
-    :param end:
-        A :class:`relativedelta.relativedelta` object or equivalent representing
-        the time and time of year that daylight savings time ends, with the
-        same specification method as in ``start``. One note is that this should
-        point to the first time in the *standard* zone, so if a transition
-        occurs at 2AM in the DST zone and the clocks are set back 1 hour to 1AM,
-        set the `hours` parameter to +1.
-
-
-    **Examples:**
-
-    .. testsetup:: tzrange
-
-        from dateutil.tz import tzrange, tzstr
-
-    .. doctest:: tzrange
-
-        >>> tzstr('EST5EDT') == tzrange("EST", -18000, "EDT")
-        True
-
-        >>> from dateutil.relativedelta import *
-        >>> range1 = tzrange("EST", -18000, "EDT")
-        >>> range2 = tzrange("EST", -18000, "EDT", -14400,
-        ...                  relativedelta(hours=+2, month=4, day=1,
-        ...                                weekday=SU(+1)),
-        ...                  relativedelta(hours=+1, month=10, day=31,
-        ...                                weekday=SU(-1)))
-        >>> tzstr('EST5EDT') == range1 == range2
-        True
-
+    A fully initialized ``tzrangebase`` subclass should also provide the
+    following attributes:
+        * ``hasdst``: Boolean whether or not the zone uses DST.
+        * ``_dst_offset`` / ``_std_offset``: :class:`datetime.timedelta` objects
+          representing the respective UTC offsets.
+        * ``_dst_abbr`` / ``_std_abbr``: Strings representing the timezone short
+          abbreviations in DST and STD, respectively.
+        * ``_hasdst``: Whether or not the zone has DST.
     """
-    def __init__(self, stdabbr, stdoffset=None,
-                 dstabbr=None, dstoffset=None,
-                 start=None, end=None):
-        super(tzrange, self).__init__()
-
-        global relativedelta
-        from dateutil import relativedelta
-
-        self._std_abbr = stdabbr
-        self._dst_abbr = dstabbr
-
-        try:
-            stdoffset = _total_seconds(stdoffset)
-        except (TypeError, AttributeError):
-            pass
-
-        try:
-            dstoffset = _total_seconds(dstoffset)
-        except (TypeError, AttributeError):
-            pass
-
-        if stdoffset is not None:
-            self._std_offset = timedelta(seconds=stdoffset)
-        else:
-            self._std_offset = ZERO
-
-        if dstoffset is not None:
-            self._dst_offset = timedelta(seconds=dstoffset)
-        elif dstabbr and stdoffset is not None:
-            self._dst_offset = self._std_offset+timedelta(hours=+1)
-        else:
-            self._dst_offset = ZERO
-
-        if dstabbr and start is None:
-            self._start_delta = relativedelta.relativedelta(
-                hours=+2, month=4, day=1, weekday=relativedelta.SU(+1))
-        else:
-            self._start_delta = start
-
-        if dstabbr and end is None:
-            self._end_delta = relativedelta.relativedelta(
-                hours=+1, month=10, day=31, weekday=relativedelta.SU(-1))
-        else:
-            self._end_delta = end
-
-        self._dst_base_offset = self._dst_offset - self._std_offset
+    def __init__(self):
+        raise NotImplementedError('tzrangebase is an abstract base class')
 
     def utcoffset(self, dt):
-        if dt is None:
-            return None
+        isdst = self._isdst(dt)
 
-        if self._isdst(dt):
+        if isdst is None:
+            return None
+        elif isdst:
             return self._dst_offset
         else:
             return self._std_offset
 
     def dst(self, dt):
-        if self._isdst(dt):
-            return self._dst_offset - self._std_offset
+        isdst = self._isdst(dt)
+
+        if isdst is None:
+            return None
+        elif isdst:
+            return self._dst_base_offset
         else:
             return ZERO
 
@@ -324,7 +228,7 @@ class tzrange(_tzinfo):
             raise ValueError("dt.tzinfo is not self")
 
         # Get transitions - if there are none, fixed offset
-        transitions = self._transitions(dt.year)
+        transitions = self.transitions(dt.year)
         if transitions is None:
             return dt + self.utcoffset(dt)
 
@@ -350,17 +254,21 @@ class tzrange(_tzinfo):
         return enfold(dt_wall, fold=_fold)
 
     def is_ambiguous(self, dt):
-        transitions = self._transitions(dt.year)
-        if transitions is None:
+        if not self.hasdst:
             return False
 
-        start, end = transitions
+        start, end = self.transitions(dt.year)
 
         dt = dt.replace(tzinfo=None)
         return (end <= dt < end + self._dst_base_offset)
 
     def _isdst(self, dt):
-        transitions = self._transitions(dt.year)
+        if not self.hasdst:
+            return False
+        elif dt is None:
+            return None
+
+        transitions = self.transitions(dt.year)
 
         if transitions is None:
             return False
@@ -387,28 +295,10 @@ class tzrange(_tzinfo):
 
         return isdst
 
-    def _transitions(self, year):
-        if not self._start_delta:
-            return None
-
-        base_year = datetime(year, 1, 1)
-
-        start = base_year + self._start_delta
-        end = base_year + self._end_delta
-
-        return (start, end)
-
-    def __eq__(self, other):
-        if not isinstance(other, tzrange):
-            return NotImplemented
-
-        return (self._std_abbr == other._std_abbr and
-                self._dst_abbr == other._dst_abbr and
-                self._std_offset == other._std_offset and
-                self._dst_offset == other._dst_offset and
-                self._start_delta == other._start_delta and
-                self._end_delta == other._end_delta)
-
+    @property
+    def _dst_base_offset(self):
+        return self._dst_offset - self._std_offset
+    
     __hash__ = None
 
     def __ne__(self, other):
